bookmarker/requirement.txt
coverage==4.0.3
Django==1.9.1
django-cors-headers==1.1.0
django-nose==1.4.3
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
gunicorn==19.4.5
mysqlclient==1.3.7
nose==1.3.7
Pillow==3.1.0
PyJWT==1.4.0
wheel==0.26.0
